Update: Modularise BorderedWell component
