+++ src/components/alexandria/BorderedWell/styles.scss
@@ -1,5 +1,5 @@
-@import '../color';
-@import '../border';
+@import 'styles/color';
+@import 'styles/border';
 
 .borderedwell-component {
   background-color: $color-white;
