+++ src/components/core.js
@@ -39,7 +39,7 @@ import fastStatelessWrapper from 'components/adslotUi/fastStatelessWrapper';
 // Alexandria
 import Alert from 'components/alexandria/Alert/component';
 import Avatar from 'components/alexandria/Avatar/component';
-import BorderedWell from 'components/alexandria/BorderedWellComponent';
+import BorderedWell from 'components/alexandria/BorderedWell/component';
 import Breadcrumb from 'components/alexandria/BreadcrumbComponent';
 import Card from 'components/alexandria/CardComponent';
 import Empty from 'components/alexandria/EmptyComponent';
